refactor: rename daily details info component to section for clarity
