+++ lib/presentation/screen/component/daily_details_info_section.dart
